# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The clinic's API",
        "version": "1"
    },
    "servers": [
        {
            "url": "https://localhost",
            "description": "Optional server description, e.g. Main (production) server"
        }
    ],
...
    "paths": {
        "/isAuthenticated": {
            "get": {
                "operationId": "is-authenticated",
                "tags": [
                    "General"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "authenticated": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/theme": {
            "get": {
                "description": "Gets the current theme that is stored in the users's session.",
                "tags": [
                    "General"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "theme": {
                                            "type": "string",
                                            "example": "dark"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            },
            "post": {
                "description": "Changes the current theme",
                "tags": [
                    "General"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "theme"
                                ],
                                "properties": {
                                    "theme": {
                                        "type": "string",
                                        "example": "dark"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "A massage to confirm the theme change."
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/General/theme.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/locale": {
            "get": {
                "description": "Gets the current locale option that is stored in the users's session.",
                "tags": [
                    "General"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "longName": {
                                            "type": "string",
                                            "example": "English"
                                        },
                                        "shortName": {
                                            "type": "string",
                                            "example": "en"
                                        },
                                        "direction": {
                                            "type": "string",
                                            "description": "The direction in which the language is written.",
                                            "example": "ltr"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            },
            "put": {
                "description": "Changes the current locale option that is stored in the users's session.",
                "tags": [
                    "General"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "locale"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "description": "Either short name or long name of the dsired locale.",
                                        "example": "en"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "A massage to confirm the locale change."
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/General/locale.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/locales": {
            "get": {
                "description": "Gets all the locale options available.",
                "tags": [
                    "General"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "longName": {
                                                "type": "string",
                                                "example": "English"
                                            },
                                            "shortName": {
                                                "type": "string",
                                                "example": "en"
                                            },
                                            "direction": {
                                                "type": "string",
                                                "description": "The direction in which the language is written.",
                                                "example": "ltr"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object",
                                        "required": [
                                            "password",
                                            "username"
                                        ],
                                        "properties": {
                                            "username": {
                                                "type": "string",
                                                "example": "fooname",
                                                "format": "username"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password"
                                            },
                                            "remember": {
                                                "type": "boolean"
                                            }
                                        }
                                    },
                                    {
                                        "type": "object",
                                        "required": [
                                            "password",
                                            "email"
                                        ],
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "example@example.example",
                                                "format": "email"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password"
                                            },
                                            "remember": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Successful, redirects the user to the home page."
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/login.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/auth/verify-phonenumber": {
            "post": {
                "description": "Recieves phonenumber and the code that has been already sent to the user and verifies them.",
                "tags": [
                    "Accounts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "pattern": "^[0-9]{6}$"
                                    },
                                    "phonenumber": {
                                        "type": "string",
                                        "pattern": "^09[0-9]{9}$",
                                        "format": "phonenumber"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "message"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/auth-verify-phonenumber.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/auth/send-code-to-phonenumber": {
            "post": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "phonenumber": {
                                        "type": "string",
                                        "pattern": "^09[0-9]{9}$",
                                        "format": "phonenumber"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "message"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/auth-send-code-to-phonenumber.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/auth/send-code-to-email": {
            "post": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "message"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/auth-send-code-to-email.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/auth/reset-password": {
            "put": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "requestBody": {
                    "content": {
                        "applcation/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "pattern": "^[0-9]{6}$"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "code expiration or mismatch",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "message"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/auth-reset-password.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/isEmailVerified": {
            "get": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "applicaiton/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "verified": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/email/verification-notification": {
            "post": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "responses": {
                    "301": {
                        "description": "redirects user to '/' or to previous url"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/logout": {
            "get": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "responses": {
                    "301": {
                        "description": "redirects user to '/' or to previous url"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
    }
}